added jwt clamis to delete endpoint

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from models.item import ItemModel
 
 class Item(Resource):
@@ -41,6 +41,9 @@ class Item(Resource):
 
     @jwt_required()
     def delete(self,name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {"message": "Admin privilege required."}, 401
         item = ItemModel.get_by_name(name)
         if item:
             item.delete_from_db()
